breaks down InvariantCheck to multiple classes
